Fix bugs with lambda scaling and unstandardized mean calculation.
Before, if scale.weights conditional at line 104 evaluated to TRUE, 'lmbds' was
not initialized. Now, 'contrasts' assigned to 'lmbds' before evaluation of the
conditional occurs.

Before, the code at line 151 exceeded 80 characters. Now, it does not.

Before, the unstandardized contrast option incorrectly used the meanDiffBoot
function. Now, it uses calcUnstandardizedMean.

Before, the error message at line 207 after the large conditional statement for
choosing a bootstrap function did not provide the user with information on the
parameters that had been passed to the function.

Before, there were no test cases for bootES with unstandardized contrasts. Now,
there is a test case for an unstandardized contrast with and without weight
scaling.