Allow user to subset the evaluation to only the groups specified 'contrasts'.
Before, all groups in the data.frame had to be present as named arguments to
'contrasts' or bootES would throw an error. Now, bootES only throws an error if
a contrast name does not correspond to a valid group. If a group is omitted in
'contrasts' but present in 'dat', the omitted group will not be included in the
calculation of the test statistic.