Simplified configuration for calculating slope.
Previously, effect.type='slope' and slope.levels had to be passed to calculate slope. Now, effect.type="slope" never
needs to be set, only slope.levels. If the user has the slope levels in the levels in the data frame as integer or real
values then she can pass the column containing these values as the 'slope.levels' parameter.